Captcha_generator/requirements.txt
Pillow==8.3.1
python==3.9